api/pyproject.toml
[project]
name = "burstdb-api"
version = "2.0.0"
description = "BurstDB — synthetic data generation API for SQL, MongoDB, and Neo4j"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "loguru>=0.7.3",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.20",
    "redis-py-cluster>=1.3.5",
    "redis[hiredis]>=6.2.0",
    "sdv>=1.11.0",
    "pandas>=2.2.0",
    "slowapi>=0.1.9",
    "sqlparse>=0.5.3",
    "supabase>=2.17.0",
    "uvicorn>=0.35.0",
    "pymongo>=4.6.0",
    "neo4j>=5.17.0",
    "httpx>=0.28.1",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "pyarrow>=15.0.0",
    "litellm>=1.20.0",
    "dspy-ai>=2.4.0",
    "torch>=2.2.0",
    "ctgan>=0.9.1",
    "sdv>=1.11.0",
    "sqlalchemy>=2.0.48",
    "asyncpg>=0.31.0",
    "boto3>=1.42.63",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.0",
    "pytest>=8.4.1",
]
